backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
Pillow==10.1.0
numpy==1.24.3
pandas==2.1.0
scikit-learn==1.3.0
matplotlib==3.8.0
seaborn==0.13.0
nltk==3.8.1
gunicorn==21.2.0
scipy==1.11.4